refactor: Entity, Repository 도메인 정리 #15
