+++ src/main/java/com/hogu/am_i_hogu/domain/User/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.hogu.am_i_hogu.domain.auth.repository;
+package com.hogu.am_i_hogu.domain.User.repository;
 
-import com.hogu.am_i_hogu.domain.auth.domain.User;
+import com.hogu.am_i_hogu.domain.User.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
