+++ src/main/java/com/hogu/am_i_hogu/domain/auth/service/AuthService.java
@@ -6,16 +6,16 @@ import com.hogu.am_i_hogu.common.exception.ErrorResponse;
 import com.hogu.am_i_hogu.common.security.JwtProvider;
 import com.hogu.am_i_hogu.common.security.TokenHasher;
 import com.hogu.am_i_hogu.common.util.TsidGenerator;
-import com.hogu.am_i_hogu.domain.oauth.domain.RefreshToken;
-import com.hogu.am_i_hogu.domain.oauth.domain.RegisterSession;
+import com.hogu.am_i_hogu.domain.auth.domain.RefreshToken;
+import com.hogu.am_i_hogu.domain.auth.domain.RegisterSession;
 import com.hogu.am_i_hogu.domain.oauth.domain.SocialAccount;
-import com.hogu.am_i_hogu.domain.auth.domain.User;
+import com.hogu.am_i_hogu.domain.User.domain.User;
 import com.hogu.am_i_hogu.domain.auth.dto.response.OnboardingResult;
 import com.hogu.am_i_hogu.domain.auth.exception.AuthErrorCode;
-import com.hogu.am_i_hogu.domain.oauth.repository.RefreshTokenRepository;
-import com.hogu.am_i_hogu.domain.oauth.repository.RegisterSessionRepository;
+import com.hogu.am_i_hogu.domain.auth.repository.RefreshTokenRepository;
+import com.hogu.am_i_hogu.domain.auth.repository.RegisterSessionRepository;
 import com.hogu.am_i_hogu.domain.oauth.repository.SocialAccountRepository;
-import com.hogu.am_i_hogu.domain.auth.repository.UserRepository;
+import com.hogu.am_i_hogu.domain.User.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
