+++ src/test/java/com/hogu/am_i_hogu/domain/oauth/OAuthServiceTest.java
@@ -7,15 +7,15 @@ import com.hogu.am_i_hogu.common.util.TsidGenerator;
 import com.hogu.am_i_hogu.domain.oauth.config.GoogleOAuthProperties;
 import com.hogu.am_i_hogu.domain.oauth.domain.OAuthLoginState;
 import com.hogu.am_i_hogu.domain.oauth.domain.OAuthProvider;
-import com.hogu.am_i_hogu.domain.oauth.domain.RefreshToken;
-import com.hogu.am_i_hogu.domain.oauth.domain.RegisterSession;
+import com.hogu.am_i_hogu.domain.auth.domain.RefreshToken;
+import com.hogu.am_i_hogu.domain.auth.domain.RegisterSession;
 import com.hogu.am_i_hogu.domain.oauth.domain.SocialAccount;
 import com.hogu.am_i_hogu.domain.oauth.dto.OAuthUserInfo;
 import com.hogu.am_i_hogu.domain.oauth.dto.response.OAuthCallbackResult;
 import com.hogu.am_i_hogu.domain.oauth.exception.OAuthErrorCode;
 import com.hogu.am_i_hogu.domain.oauth.repository.OAuthLoginStateRepository;
-import com.hogu.am_i_hogu.domain.oauth.repository.RefreshTokenRepository;
-import com.hogu.am_i_hogu.domain.oauth.repository.RegisterSessionRepository;
+import com.hogu.am_i_hogu.domain.auth.repository.RefreshTokenRepository;
+import com.hogu.am_i_hogu.domain.auth.repository.RegisterSessionRepository;
 import com.hogu.am_i_hogu.domain.oauth.repository.SocialAccountRepository;
 import com.hogu.am_i_hogu.domain.oauth.service.OAuthCallbackHandler;
 import com.hogu.am_i_hogu.domain.oauth.service.OAuthCallbackHandlerFactory;
